+++ Editor/GameObjectFavoriteIconType.cs
@@ -1,0 +1,15 @@
+using System;
+using UnityEngine;
+
+namespace SaintsHierarchy.Editor
+{
+    [Serializable]
+    public enum GameObjectFavoriteIconType
+    {
+        [InspectorName("Default (Not Supported Yet)")]
+        Default,
+        UnityDefault,
+        None,
+        Custom,
+    }
+}