+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.1 ##
+
+1.  Add: You can now change alias and icon for a favorite gameObject
+2.  Improve: Better dragging detection for favorite gameObject
+
 ## 1.2.0 ##
 
 Add: You can now drag & drop GameObjects to Favorite area at the top of the hierarchy.